requirements.txt
amqp==1.4.9
anyjson==0.3.3
beautifulsoup4==4.5.1
billiard==3.3.0.23
boto==2.42.0
celery==3.1.23
charade==1.0.3
cssselect==0.9.2
cycler==0.10.0
decorator==4.0.10
Django==1.9.8
django-cors-headers==1.1.0
djangorestframework==3.4.6
elasticsearch==2.4.0
jieba==0.38
jsonfield==1.0.3
kombu==3.0.35
lxml==3.6.4
mysqlclient==1.3.7
networkx==1.11
nltk==3.2.1
numpy==1.11.1
PyMySQL==0.7.9
pyparsing==2.1.9
python-dateutil==2.5.3
pytz==2016.6.1
redis==2.10.5
requests==2.11.1
shortuuid==0.4.3
six==1.10.0
textrank4zh==0.2
url-normalize==1.3.1
urllib3==1.16
uuid==1.30
w3lib==1.15.0
